rename workflow page
